+++ CHANGELOG.rst
@@ -10,7 +10,7 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
 **Added**
 
-* Introduced new generic interface ``life.qbic.utils.DataSetParser`` that can be used to
+* Introduced new generic interface ``life.qbic.datasets.parsers.DataSetParser`` that can be used to
 implement dataset parsers for different dataset types.
 
 **Fixed**